# Use `ureq` for "science" download, not `cargo install`'d `ptex`