--- a/package/Cargo.toml
+++ b/package/Cargo.toml
@@ -15,12 +15,12 @@ dirs = "5.0"
 fd-lock = "4.0"
 lazy_static = "1.4"
 log = { workspace = true }
-once_cell = "1.19"
 pretty_env_logger = "0.5"
 regex = "1.10"
 serde_json = "1.0.114"
 sha2 = "0.10"
 tempfile = { workspace = true }
 termcolor = "1.4"
+ureq = "2.9"
 url = "2.5"
 walkdir = "2.4"