--- a/package/Cargo.toml
+++ b/package/Cargo.toml
@@ -22,5 +22,4 @@ sha2 = "0.10"
 tempfile = { workspace = true }
 termcolor = "1.4"
 ureq = "2.9"
-url = "2.5"
 walkdir = "2.4"