@huonw 2024-04-24
I did a CI build without the cache, because the cached build wasn't exercising this new code path. Seems to work ✅